use double instead of float in preferences
git-svn-id: https://bibdesk.svn.sourceforge.net/svnroot/bibdesk/trunk/bibdesk@17922 0fceea05-a30d-0410-8a8b-80ef821fb0a1

+++ BibDocument_UI.m
@@ -1262,12 +1262,12 @@ static void applyChangesToCiteFieldsWithInfo(const void *citeField, void *contex
         if (object == sideFileView) {
             FVDisplayMode displayMode = [sideFileView displayMode];
             CGFloat iconScale = displayMode == FVDisplayModeGrid ? [sideFileView iconScale] : 0.0;
-            [[NSUserDefaults standardUserDefaults] setFloat:iconScale forKey:BDSKSideFileViewIconScaleKey];
+            [[NSUserDefaults standardUserDefaults] setDouble:iconScale forKey:BDSKSideFileViewIconScaleKey];
             [[NSUserDefaults standardUserDefaults] setInteger:displayMode forKey:BDSKSideFileViewDisplayModeKey];
         } else if (object == bottomFileView) {
             FVDisplayMode displayMode = [bottomFileView displayMode];
             CGFloat iconScale = displayMode == FVDisplayModeGrid ? [bottomFileView iconScale] : 0.0;
-            [[NSUserDefaults standardUserDefaults] setFloat:iconScale forKey:BDSKBottomFileViewIconScaleKey];
+            [[NSUserDefaults standardUserDefaults] setDouble:iconScale forKey:BDSKBottomFileViewIconScaleKey];
             [[NSUserDefaults standardUserDefaults] setInteger:displayMode forKey:BDSKBottomFileViewDisplayModeKey];
         }
     } else if (context == &BDSKDocumentDefaultsObservationContext) {
